Move library render into own view
